Get rid of bufsize adjustment logic
Currently, this script tries to adapt to the host it's running on by
checking the return value of send() to see if each segment buffer was
fully sent and downsizing the buffer if not. Since the goal is to have
only one segment in flight at a time, this is supposed to detect when
the kernel sends the buffer as multiple segments and fix it for the next
buffer.

However, this doesn't work in practice because send() will always accept
as much data as it can buffer in a single call, even if that data ends
up as multiple segments on the wire. Since we set the kernel send buffer
size to our buffer size, that means the send() call will always accept
the full buffer and the downsizing logic will never run at all!

Since the code doesn't run, get rid of it and update the comments to
better reflect the kernel's actual behavior.

It's possible there's some kernel feature or Linux-specific sysctl we
could use to accurately query the size of sent segments or that there's
a similar feature (maybe TCP_CORK) we could use to force the segment
size we want. Those changes would be more invasive to the existing code,
though, and it's probably not worth the time given how few of these
devices are likely around anymore.